feat: move admin pages to admin module
