+++ packages/admin/package.json
@@ -31,14 +31,17 @@
         "test:watch": "vitest watch"
     },
     "dependencies": {
+        "bootstrap-vue-next": "^0.18.1",
+        "@authup/core-kit": "^1.0.0-beta.13",
         "@authup/client-web-kit": "^1.0.0-beta.15",
         "@dnpm-dip/core": "^1.11.0",
         "@dnpm-dip/kit": "^1.10.0",
-        "@nuxt/kit": "^3.11.1"
+        "@nuxt/kit": "^3.11.1",
+        "@vuecs/timeago": "^1.1.0",
+        "vue": "^3.4.27"
     },
     "devDependencies": {
         "@nuxt/module-builder": "^0.6.0",
-        "nuxt": "^3.11.2",
-        "vue": "^3.4.27"
+        "nuxt": "^3.11.2"
     }
 }
