+++ packages/admin/src/runtime/pages/identity-providers/[id]/index.vue
