+++ packages/admin/src/runtime/pages/identity-providers/index.vue
@@ -1,9 +1,17 @@
+<!--
+  - Copyright (c) 2024.
+  - Author Peter Placzek (tada5hi)
+  - For the full copyright and license information,
+  - view the LICENSE file that was distributed with this source code.
+  -->
+
 <script lang="ts">
+
+import type { IdentityProvider } from '@authup/core-kit';
+import { PermissionName } from '@authup/core-kit';
 import {
     DNav, PageMetaKey, PageNavigationTopID, useToast,
 } from '@dnpm-dip/core';
-import type { User } from '@authup/core-kit';
-import { PermissionName } from '@authup/core-kit';
 import { definePageMeta } from '#imports';
 import { defineNuxtComponent } from '#app';
 
@@ -13,12 +21,12 @@ export default defineNuxtComponent({
     },
     setup() {
         definePageMeta({
-            [PageMetaKey.NAVIGATION_TOP_ID]: PageNavigationTopID.ADMIN,
             [PageMetaKey.REQUIRED_LOGGED_IN]: true,
+            [PageMetaKey.NAVIGATION_TOP_ID]: PageNavigationTopID.ADMIN,
             [PageMetaKey.REQUIRED_PERMISSIONS]: [
-                PermissionName.USER_EDIT,
-                PermissionName.USER_ADD,
-                PermissionName.USER_DROP,
+                PermissionName.PROVIDER_EDIT,
+                PermissionName.PROVIDER_DROP,
+                PermissionName.PROVIDER_ADD,
             ],
         });
 
@@ -32,14 +40,14 @@ export default defineNuxtComponent({
             },
             {
                 name: 'Hinzufügen',
-                urlSuffix: '/add',
+                urlSuffix: 'add',
                 icon: 'fa fa-plus',
             },
         ];
 
-        const handleDeleted = (e: User) => {
+        const handleDeleted = (e: IdentityProvider) => {
             if (toast) {
-                toast.show({ variant: 'success', body: `The user ${e.name} was successfully deleted.` });
+                toast.show({ variant: 'success', body: `The identity-provider ${e.name} was successfully deleted.` });
             }
         };
 
@@ -50,9 +58,9 @@ export default defineNuxtComponent({
         };
 
         return {
-            items,
-            handleFailed,
             handleDeleted,
+            handleFailed,
+            items,
         };
     },
 });
@@ -60,16 +68,14 @@ export default defineNuxtComponent({
 <template>
     <div>
         <h1 class="title no-border mb-3">
-            <i class="fa fa-user me-1" /> Benutzer
-            <span class="sub-title ms-1">
-                Verwaltung
-            </span>
+            <i class="fa-solid fa-atom me-1" /> Identitätsanbieter
+            <span class="sub-title ms-1">Verwaltung</span>
         </h1>
         <div class="content-wrapper">
             <div class="content-sidebar flex-column">
                 <DNav
                     :items="items"
-                    path="/admin/users"
+                    path="/admin/identity-providers"
                     direction="vertical"
                 />
             </div>
