+++ packages/admin/src/runtime/pages/identity-providers/index/index.vue
@@ -1,15 +1,15 @@
 <script lang="ts">
+
+import { VCTimeago } from '@vuecs/timeago';
 import { BTable } from 'bootstrap-vue-next';
-import type { User } from '@authup/core-kit';
+import type { IdentityProvider } from '@authup/core-kit';
+import { PermissionName, isRealmResourceWritable } from '@authup/core-kit';
 import {
-    PermissionName, isRealmResourceWritable,
-} from '@authup/core-kit';
-import {
-    AEntityDelete, APagination, ASearch, ATitle, AUsers, useAbilityCheck, useStore,
+    AEntityDelete, AIdentityProviders, APagination, ASearch, ATitle, storeToRefs, useAbilityCheck,
+    useStore,
 } from '@authup/client-web-kit';
-import { storeToRefs } from 'pinia';
 import type { BuildInput } from 'rapiq';
-import { defineNuxtComponent } from '#imports';
+import { defineNuxtComponent } from '#app';
 
 export default defineNuxtComponent({
     components: {
@@ -17,30 +17,31 @@ export default defineNuxtComponent({
         APagination,
         ASearch,
         BTable,
-        AUsers,
+        AIdentityProviders,
         AEntityDelete,
+        VCTimeago,
     },
     emits: ['deleted'],
     setup(props, { emit }) {
-        const handleDeleted = (e: User) => {
+        const handleDeleted = (e: IdentityProvider) => {
             emit('deleted', e);
         };
 
         const store = useStore();
         const { realm, realmManagementId } = storeToRefs(store);
 
-        const query : BuildInput<User> = {
+        const query : BuildInput<IdentityProvider> = {
             filter: {
                 realm_id: [realmManagementId.value, null],
             },
         };
 
         const isResourceWritable = (
-            resource: User,
+            resource: IdentityProvider,
         ) => isRealmResourceWritable(realm.value, resource.realm_id);
 
-        const hasEditPermission = useAbilityCheck(PermissionName.USER_EDIT);
-        const hasDropPermission = useAbilityCheck(PermissionName.USER_DROP);
+        const hasEditPermission = useAbilityCheck(PermissionName.PROVIDER_EDIT);
+        const hasDropPermission = useAbilityCheck(PermissionName.PROVIDER_DROP);
 
         const fields = [
             {
@@ -55,7 +56,9 @@ export default defineNuxtComponent({
             {
                 key: 'updated_at', label: 'Aktualisierungsdatum', thClass: 'text-left', tdClass: 'text-left',
             },
-            { key: 'options', label: '', tdClass: 'text-left' },
+            {
+                key: 'options', label: '', tdClass: 'text-left',
+            },
         ];
 
         return {
@@ -70,10 +73,8 @@ export default defineNuxtComponent({
 });
 </script>
 <template>
-    <AUsers
+    <AIdentityProviders
         :query="query"
-        :body="{tag: 'div'}"
-        :footer="true"
         @deleted="handleDeleted"
     >
         <template #header="props">
@@ -106,7 +107,7 @@ export default defineNuxtComponent({
                 </template>
                 <template #cell(options)="data">
                     <NuxtLink
-                        :to="'/admin/users/'+ data.item.id"
+                        :to="'/admin/identity-providers/'+ data.item.id"
                         class="btn btn-xs btn-outline-primary me-1"
                         :disabled="!hasEditPermission || !isResourceWritable(data.item)"
                     >
@@ -115,7 +116,7 @@ export default defineNuxtComponent({
                     <AEntityDelete
                         class="btn btn-xs btn-outline-danger"
                         :entity-id="data.item.id"
-                        entity-type="user"
+                        entity-type="identityProvider"
                         :with-text="false"
                         :disabled="!hasDropPermission || !isResourceWritable(data.item)"
                         @deleted="props.deleted"
@@ -123,5 +124,5 @@ export default defineNuxtComponent({
                 </template>
             </BTable>
         </template>
-    </AUsers>
+    </AIdentityProviders>
 </template>
