+++ packages/admin/src/runtime/pages/roles/index/add.vue
@@ -1,9 +1,8 @@
 <script lang="ts">
-import { ARoleForm, useStore } from '@authup/client-web-kit';
+import { ARoleForm, storeToRefs, useStore } from '@authup/client-web-kit';
 import type { Role } from '@authup/core-kit';
 import { PermissionName } from '@authup/core-kit';
 import { PageMetaKey, PageNavigationTopID } from '@dnpm-dip/core';
-import { storeToRefs } from 'pinia';
 import { defineNuxtComponent, navigateTo } from '#app';
 import { definePageMeta } from '#imports';
